+++ bootstrap/app.js
@@ -1,6 +1,9 @@
-import { ApolloServer } from 'apollo-server'
+import Koa from 'koa'
+import { ApolloServer } from 'apollo-server-koa'
 import { typeDefs, resolvers } from '../schema'
+import onHealthCheck from '../middlewares/healthCheck'
 
+const app = new Koa()
 const { PORT } = process.env
 
 const server = new ApolloServer({
@@ -8,8 +11,11 @@ const server = new ApolloServer({
   resolvers,
 })
 
-const app = server.listen(PORT).then(({ url }) => {
-  console.log(`🚀 appication running at ${url}`) // eslint-disable-line no-console
+server.applyMiddleware({
+  app,
+  onHealthCheck,
 })
 
-export default app
+export default app.listen({ port: PORT }, () => {
+  console.log(`🚀 appication running at port ${PORT}`) // eslint-disable-line no-console
+})