+++ bootstrap/database.js
@@ -10,11 +10,10 @@ const {
 
 const uri = `mongodb://${DB_HOST}:${DB_PORT}/${DB_NAME}`
 
-export const connect = () =>
-  mongoose.connect(uri, {
-    user: DB_USER,
-    pass: DB_PASS,
-    useNewUrlParser: true,
-  })
+export const connect = mongoose.connect(uri, {
+  user: DB_USER,
+  pass: DB_PASS,
+  useNewUrlParser: true,
+})
 
 export const { connection } = mongoose